backend/requirements.txt
# - bcrypt: Password hashing
# - PyJWT: JWT token generation and validation

fastapi
uvicorn
pydantic
sqlalchemy
pyodbc
python-dotenv
schedule
pytz
jinja2
cryptography
bcrypt
PyJWT